refactor: use a single type file and require names lookup in hexToName
