+++ src/cmykToRgb.ts
@@ -1,16 +1,15 @@
 import { parseCmyk } from './parseCmyk';
-import { ICmykColor } from './types/cmyk';
-import { IRgbColor } from './types/rgb';
+import { ICmykColor, IRgbColor } from './types';
 import { normalizeDecNum } from './utils';
 
 export function cmykToRgb (cmyk: ICmykColor | string | null): IRgbColor | null {
-  cmyk = typeof cmyk === 'string' ? parseCmyk(cmyk) : cmyk;
-  if (!cmyk) {
+  const parsed = typeof cmyk === 'string' ? parseCmyk(cmyk) : cmyk;
+  if (!parsed) {
     return null;
   }
 
   // Convert the CMYK values to the range 0-1
-  const [cyan, magenta, yellow, key, alpha] = [cmyk.cyan / 100, cmyk.magenta / 100, cmyk.yellow / 100, cmyk.key / 100, cmyk.alpha];
+  const [cyan, magenta, yellow, key, alpha] = [parsed.cyan / 100, parsed.magenta / 100, parsed.yellow / 100, parsed.key / 100, parsed.alpha];
 
   // Calculate the rgb values
   const red = normalizeDecNum(255 * (1 - cyan) * (1 - key));
