+++ src/isDark.ts
@@ -1,6 +1,5 @@
 import { relativeLuminance } from "./relativeLuminance";
-import { IRgbColor } from "./types/rgb";
-
+import { IRgbColor } from "./types";
 
 export function isDark(value: IRgbColor): boolean {
   const L = relativeLuminance(value);
