+++ src/isValidHex.ts
@@ -1,5 +1,5 @@
 import { parseHex } from './parseHex';
-import { IHexColor } from './types/hex';
+import { IHexColor } from './types';
 
 export function isValidHex (value: IHexColor | string | null) {
   const normalizedVal = typeof value === 'string' ? parseHex(value) : value;
