+++ src/parseHsv.ts
@@ -1,4 +1,4 @@
-import { IHsvColor } from './types/hsv';
+import { IHsvColor } from './types';
 
 export interface IHsvParsingOptions {
   allowDecimal: boolean;
