--- src/types/cmyk.ts
@@ -1,7 +1,0 @@
-export interface ICmykColor {
-  alpha?: number;
-  cyan: number;
-  magenta: number;
-  yellow: number;
-  key: number;
-}
