--- src/types/hex.ts
@@ -1,6 +1,0 @@
-export interface IHexColor {
-  alpha?: string;
-  red: string;
-  green: string;
-  blue: string;
-}
