--- src/types/hsl.ts
@@ -1,6 +1,0 @@
-export interface IHslColor {
-  hue: number;
-  lum: number;
-  sat: number;
-  alpha?: number;
-}
