--- src/types/hsv.ts
@@ -1,6 +1,0 @@
-export interface IHsvColor {
-  hue: number;
-  val: number;
-  sat: number;
-  alpha?: number;
-};
