--- src/types/rgb.ts
@@ -1,6 +1,0 @@
-export interface IRgbColor {
-  alpha?: number;
-  red: number;
-  green: number;
-  blue: number;
-}
